minor change to db in telegram bot

+++ telegram_bot/bd_classes/tables/PointOnTheMap.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, Float
+from sqlalchemy import Column, BigInteger, String, ForeignKey, Float
 from sqlalchemy.orm import relationship
 
 from bd_classes.tables.Base import Base
@@ -8,12 +8,12 @@ class PointOnTheMapTable(Base):
     __tablename__ = 'point_on_the_map'
     __table_args__ = {'schema': 'public'}
 
-    message_id = Column(Integer, primary_key=True, unique=True, nullable=False)
+    message_id = Column(BigInteger, primary_key=True, unique=True, nullable=False)
     photo_path = Column(String)
     latitude = Column(Float, nullable=False)
     longitude = Column(Float, nullable=False)
     message = Column(String)
-    user_id = Column(Integer, ForeignKey('public.user.user_id'), nullable=False)
+    user_id = Column(BigInteger, ForeignKey('public.user.user_id'), nullable=False)
 
     user_relationship = relationship("UserTable")
